backend/test-requirements.txt
pytest==5.2.2
pytest-django==3.6.0
